+++ .pnp.cjs
@@ -487,6 +487,10 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:packages/sources/metalsapi"
       },
       {
+        "name": "@chainlink/mock-ea-adapter",
+        "reference": "workspace:packages/sources/mock-ea"
+      },
+      {
         "name": "@chainlink/mycryptoapi-adapter",
         "reference": "workspace:packages/sources/mycryptoapi"
       },
@@ -776,6 +780,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@chainlink/medianizer-adapter", ["workspace:packages/composites/medianizer"]],
       ["@chainlink/messari-adapter", ["workspace:packages/sources/messari"]],
       ["@chainlink/metalsapi-adapter", ["workspace:packages/sources/metalsapi"]],
+      ["@chainlink/mock-ea-adapter", ["workspace:packages/sources/mock-ea"]],
       ["@chainlink/mycryptoapi-adapter", ["workspace:packages/sources/mycryptoapi"]],
       ["@chainlink/ncfx-adapter", ["workspace:packages/sources/ncfx"]],
       ["@chainlink/nikkei-adapter", ["workspace:packages/sources/nikkei"]],
@@ -5154,6 +5159,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@chainlink/marketstack-adapter", "workspace:packages/sources/marketstack"],
             ["@chainlink/messari-adapter", "workspace:packages/sources/messari"],
             ["@chainlink/metalsapi-adapter", "workspace:packages/sources/metalsapi"],
+            ["@chainlink/mock-ea-adapter", "workspace:packages/sources/mock-ea"],
             ["@chainlink/mycryptoapi-adapter", "workspace:packages/sources/mycryptoapi"],
             ["@chainlink/ncfx-adapter", "workspace:packages/sources/ncfx"],
             ["@chainlink/nikkei-adapter", "workspace:packages/sources/nikkei"],
@@ -5333,6 +5339,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@types/jest", "npm:27.0.3"],
             ["@types/node", "npm:16.11.19"],
             ["hardhat", "npm:2.7.0"],
+            ["mock-socket", "npm:9.1.0"],
             ["tslib", "npm:2.3.1"],
             ["typescript", "patch:typescript@npm%3A4.3.5#~builtin<compat/typescript>::version=4.3.5&hash=d8b4e7"]
           ],
@@ -6153,6 +6160,22 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "linkType": "SOFT",
         }]
       ]],
+      ["@chainlink/mock-ea-adapter", [
+        ["workspace:packages/sources/mock-ea", {
+          "packageLocation": "./packages/sources/mock-ea/",
+          "packageDependencies": [
+            ["@chainlink/mock-ea-adapter", "workspace:packages/sources/mock-ea"],
+            ["@chainlink/ea-bootstrap", "workspace:packages/core/bootstrap"],
+            ["@chainlink/ea-test-helpers", "workspace:packages/core/test-helpers"],
+            ["@chainlink/types", "workspace:packages/core/types/@chainlink"],
+            ["@types/jest", "npm:27.0.3"],
+            ["@types/node", "npm:16.11.19"],
+            ["tslib", "npm:2.3.1"],
+            ["typescript", "patch:typescript@npm%3A4.3.5#~builtin<compat/typescript>::version=4.3.5&hash=d8b4e7"]
+          ],
+          "linkType": "SOFT",
+        }]
+      ]],
       ["@chainlink/mycryptoapi-adapter", [
         ["workspace:packages/sources/mycryptoapi", {
           "packageLocation": "./packages/sources/mycryptoapi/",
@@ -6182,6 +6205,8 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["@chainlink/types", "workspace:packages/core/types/@chainlink"],
             ["@types/jest", "npm:27.0.3"],
             ["@types/node", "npm:16.11.19"],
+            ["@types/supertest", "npm:2.0.11"],
+            ["supertest", "npm:6.1.6"],
             ["tslib", "npm:2.3.1"],
             ["typescript", "patch:typescript@npm%3A4.3.5#~builtin<compat/typescript>::version=4.3.5&hash=d8b4e7"]
           ],
@@ -25226,6 +25251,15 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./.yarn/cache/mock-fs-npm-4.14.0-c51913d559-dccd976a8d.zip/node_modules/mock-fs/",
           "packageDependencies": [
             ["mock-fs", "npm:4.14.0"]
+          ],
+          "linkType": "HARD",
+        }]
+      ]],
+      ["mock-socket", [
+        ["npm:9.1.0", {
+          "packageLocation": "./.yarn/cache/mock-socket-npm-9.1.0-ccd2b5bee9-19b16017da.zip/node_modules/mock-socket/",
+          "packageDependencies": [
+            ["mock-socket", "npm:9.1.0"]
           ],
           "linkType": "HARD",
         }]