+++ packages/composites/reference-transform/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/reference-transform-adapter
 
+## 1.0.45
+
+### Patch Changes
+
+- Updated dependencies [4865d3b46]
+  - @chainlink/ea-bootstrap@1.10.5
+  - @chainlink/ea@1.0.45
+  - @chainlink/ea-reference-data-reader@1.1.1
+
 ## 1.0.44
 
 ### Patch Changes