+++ packages/sources/blockcypher/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/blockcypher-adapter
 
+## 1.2.11
+
+### Patch Changes
+
+- Updated dependencies [4865d3b46]
+  - @chainlink/ea-bootstrap@1.10.5
+  - @chainlink/ea-factories@1.0.26
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.2.10
 
 ### Patch Changes