+++ packages/sources/chain-reserve-wallet/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/chain-reserve-wallet-adapter
 
+## 2.2.7
+
+### Patch Changes
+
+- b2e1df0c3: Add description for endpoints from READMEs
+- Updated dependencies [4865d3b46]
+  - @chainlink/ea-bootstrap@1.10.5
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 2.2.6
 
 ### Patch Changes