+++ packages/sources/wbtc-address-set/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/wbtc-address-set-adapter
 
+## 1.3.8
+
+### Patch Changes
+
+- Updated dependencies [4865d3b46]
+  - @chainlink/ea-bootstrap@1.10.5
+  - @chainlink/proof-of-reserves-adapter@1.4.8
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.3.7
 
 ### Patch Changes